+++ include/kernel/kernel.h
@@ -22,8 +22,10 @@
 #include <signal.h>
 #include <stdbool.h>
 #include <curses.h>
+#include <debug.h>
 
 #include <winix/type.h>
+#include <winix/page.h>
 #include <winix/comm.h>
 #include <winix/kdebug.h>
 #include <winix/bitmap.h>
@@ -39,7 +41,7 @@
 #define FREE_MEM_END    0x1ffff
 
 //Memory limits
-extern unsigned long TEXT_BEGIN, DATA_BEGIN, BSS_BEGIN;
-extern unsigned long TEXT_END, DATA_END, BSS_END;
+extern unsigned int TEXT_BEGIN, DATA_BEGIN, BSS_BEGIN;
+extern unsigned int TEXT_END, DATA_END, BSS_END;
 
 #endif
