+++ include/sys/types.h
@@ -1,16 +1,16 @@
 #ifndef _TYPE_H_
 #define _TYPE_H_ 1
 
-typedef unsigned long size_t;
+typedef unsigned int size_t;
 typedef unsigned char byte;
 typedef unsigned char byte_t;
-typedef unsigned long uint32_t;
-typedef unsigned long clock_t;
-typedef unsigned long mode_t;
+typedef unsigned int uint32_t;
+typedef unsigned int clock_t;
+typedef unsigned int mode_t;
 typedef unsigned int uid_t;
 typedef unsigned int gid_t;
-typedef unsigned long off_t;
-typedef unsigned long time_t;
+typedef unsigned int off_t;
+typedef unsigned int time_t;
 typedef unsigned int nlink_t;
 typedef int pid_t;
