+++ include/util.h
@@ -1,7 +1,7 @@
 #ifndef _UTIL_H_
 #define _UTIL_H_ 1
 
-const char *size_t_to_binary(unsigned long x);
+const char *size_t_to_binary(unsigned int x);
 int hex2int(char *a, int len);
 
 #endif
