+++ include/winix/slab.h
@@ -3,10 +3,10 @@
 
 #include <winix/gfp.h>
 
-void* kmalloc (unsigned long size);
+void* kmalloc (unsigned int size);
 void kfree(void *ptr);
-void* krealloc(void *p,unsigned long size);
-void *kcalloc(unsigned long number , unsigned long size);
+void* krealloc(void *p,unsigned int size);
+void *kcalloc(unsigned int number , unsigned int size);
 void kblock_overview();
 void init_slab(void *addr, int size);
