+++ kernel/system.c
@@ -53,7 +53,7 @@ void print_sysinfo(){
 	kprintf("Data Segment: 0x%08x - 0x%08x\r\n", &DATA_BEGIN, &DATA_END);
 	kprintf("BSS Segment:  0x%08x - 0x%08x\r\n", &BSS_BEGIN, &BSS_END);
 	kprintf("Unallocated:  0x%08x - 0x%08x\r\n", free_mem_begin, FREE_MEM_END);
-	kprintf("%d kWords Free\r\n", ((unsigned long)(FREE_MEM_END - free_mem_begin)) / 1024);
+	kprintf("%d kWords Free\r\n", ((unsigned int)(FREE_MEM_END - free_mem_begin)) / 1024);
 	print_sysmap();
 }
 /**
