+++ lib/makecontext.c
@@ -20,7 +20,7 @@ void makecontext(ucontext_t *ucp, void (* func)(), int argc, ...){
 		return;
 
 	ucp->pc = (void (*)())&_ctx_start;
-	ucp->regs[11] = (unsigned long)&_ctx_end;
+	ucp->regs[11] = (unsigned int)&_ctx_end;
 
 	//allocate stack for the ucp context
 	spp = (!ucp->ss_flags) ? &ucp->ss_sp : &ucp->sp;
